MISUV-2862 refactor package names and move Calculation into its own file
