+++ app/models/liabilitycalculation/AllowancesAndDeductions.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package models.LiabilityCalculation
+package models.liabilitycalculation
 
 import play.api.libs.json._
 import play.api.libs.functional.syntax._
