+++ app/models/liabilitycalculation/taxcalculation/TaxCalculation.scala
@@ -14,9 +14,8 @@
  * limitations under the License.
  */
 
-package models.LiabilityCalculation.TaxCalculationModels
+package models.liabilitycalculation.taxcalculation
 
-import play.api.libs.functional.syntax._
 import play.api.libs.json._
 
 case class TaxCalculation(
